+++ services/cdn/mgmt/2020-09-01/cdn/customdomains.go
@@ -32,14 +32,14 @@ type CustomDomainsClient struct {
 }
 
 // NewCustomDomainsClient creates an instance of the CustomDomainsClient client.
-func NewCustomDomainsClient(subscriptionID string, subscriptionID1 string) CustomDomainsClient {
-	return NewCustomDomainsClientWithBaseURI(DefaultBaseURI, subscriptionID, subscriptionID1)
+func NewCustomDomainsClient(subscriptionID string) CustomDomainsClient {
+	return NewCustomDomainsClientWithBaseURI(DefaultBaseURI, subscriptionID)
 }
 
 // NewCustomDomainsClientWithBaseURI creates an instance of the CustomDomainsClient client using a custom endpoint.
 // Use this when interacting with an Azure cloud that uses a non-standard base URI (sovereign clouds, Azure stack).
-func NewCustomDomainsClientWithBaseURI(baseURI string, subscriptionID string, subscriptionID1 string) CustomDomainsClient {
-	return CustomDomainsClient{NewWithBaseURI(baseURI, subscriptionID, subscriptionID1)}
+func NewCustomDomainsClientWithBaseURI(baseURI string, subscriptionID string) CustomDomainsClient {
+	return CustomDomainsClient{NewWithBaseURI(baseURI, subscriptionID)}
 }
 
 // Create creates a new custom domain within an endpoint.