+++ services/cdn/mgmt/2020-09-01/cdn/endpoints.go
@@ -32,14 +32,14 @@ type EndpointsClient struct {
 }
 
 // NewEndpointsClient creates an instance of the EndpointsClient client.
-func NewEndpointsClient(subscriptionID string, subscriptionID1 string) EndpointsClient {
-	return NewEndpointsClientWithBaseURI(DefaultBaseURI, subscriptionID, subscriptionID1)
+func NewEndpointsClient(subscriptionID string) EndpointsClient {
+	return NewEndpointsClientWithBaseURI(DefaultBaseURI, subscriptionID)
 }
 
 // NewEndpointsClientWithBaseURI creates an instance of the EndpointsClient client using a custom endpoint.  Use this
 // when interacting with an Azure cloud that uses a non-standard base URI (sovereign clouds, Azure stack).
-func NewEndpointsClientWithBaseURI(baseURI string, subscriptionID string, subscriptionID1 string) EndpointsClient {
-	return EndpointsClient{NewWithBaseURI(baseURI, subscriptionID, subscriptionID1)}
+func NewEndpointsClientWithBaseURI(baseURI string, subscriptionID string) EndpointsClient {
+	return EndpointsClient{NewWithBaseURI(baseURI, subscriptionID)}
 }
 
 // Create creates a new CDN endpoint with the specified endpoint name under the specified subscription, resource group